Refactor all the entity classes and PlayerDTO. Update the logic for both the Schedulers and add new method to update time in every minute.

+++ src/main/java/com/example/comeonplayerserviceassignment/dto/PlayerDTO.java
@@ -18,10 +18,10 @@ public class PlayerDTO
     private String surName;
     private LocalDate dateOfBirth;
     private AddressDTO address;
-    private int dailyTimeLimit = -1;
+    private long dailyTimeLimit = -1;
     private List<SessionDTO> sessions;
 
-    public PlayerDTO(long id, String email, String firstName, String surName, LocalDate dateOfBirth, AddressDTO address, int dailyTimeLimit) {
+    public PlayerDTO(long id, String email, String firstName, String surName, LocalDate dateOfBirth, AddressDTO address, long dailyTimeLimit) {
         this.id = id;
         this.email = email;
         this.firstName = firstName;
